First stab at picking up signal mapping from IInjector, if present.
Note that the behavior is now unfortunately a little schizophrenic:
if the command map's injector has a mapping for an IInjector itself,
that IInjector is used to instantiate the signal. However, this
approach ignores that (mapped) IInjector, and only looks for the
signal in the map's own IInjector. It's not immediately clear what
precedence order makes sense here, though.