Another approach to reusing IInjector mappings for signals.
Here, we follow the signal-injector logic more closely. That is,
if a mapping for an IInjector exists in the signal command map's
own injector, we use that for the signal injector (if not, we
simply use the map's injector). If the injector *does* have a
mapping, we ask it for an instance of the signal. If it does
not, we ask it to instantiate (but not map) the signal.

This is probably more sensible than only using the original
injector, but it leaves a somewhat ugly problem: if the main
injector does have a mapping for the signal, but the mapped
injector does not. We probably want to fall back to the main
injector before asking the mapped injector to instantiate
a new signal in that case.

the  injector for the signal value lookup.